requirements.txt
aenum==3.1.15
click==8.1.7
colorama==0.4.6
markdown-it-py==3.0.0
mdurl==0.1.2
prompt-toolkit==3.0.41
Pygments==2.17.2
PyYAML==6.0.1
rich==13.7.0
shellingham==1.5.4
typing_extensions==4.8.0
wcwidth==0.2.12
